Rename `utils` submodule to `util`
