+++ message_ix_buildings/chilled/postprocess/postprocess.py
@@ -2,7 +2,7 @@ import os
 
 import pandas as pd
 
-from message_ix_buildings.chilled.utils.config import Config
+from message_ix_buildings.chilled.util.config import Config
 
 cfg = Config()
