+++ message_ix_buildings/chilled/run_agg.py
@@ -3,7 +3,7 @@ import sys
 from argparse import ArgumentParser
 
 from core.climate import process_iso_tables
-from utils.config import Config  # type: ignore
+from util.config import Config  # type: ignore
 
 
 def parse_arguments(arguments):
